apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
